Created Login Modal
still need to create register modal, still need to setup with db so
that it can create and login to accounts. still need to change
scope.userState and add userState info to navbar objects like My
Account Page so that when logged in, uses actually see what they need
to.